# openapi.yaml
openapi: 3.1.0
info:
  title: Supplier Rating Software API
  description: |
    REST API interface for Supplier Rating Software (SRS).
    The backend acts as a middleware to openBIS (via JSON-RPC).
    
    The data model strictly follows the openBIS Object Types defined in the excel exports:
    - LIEFERANT (Supplier)
    - BESTELLUNG (Order)
    - BESTELLBEWERTUNG (Rating)
    
    IDs in paths refer to the openBIS **PermID** (stable technical identifier).
    Codes in payloads refer to the openBIS **Code** (human readable business identifier).
  version: 1.4.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server

paths:
  # --------------------------------------------------------------------------
  # SUPPLIERS
  # --------------------------------------------------------------------------
  /suppliers:
    get:
      summary: Retrieve list of all suppliers
      description: Returns a list of all suppliers including their calculated aggregated statistics.
      operationId: getAllSuppliers
      responses:
        '200':
          description: 'Successful retrieval [returns an array of SupplierSummaryDTO]'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SupplierSummaryDTO'
    post:
      summary: Create new supplier
      description: Creates a new supplier (Type LIEFERANT) in openBIS. The code is generated automatically.
      operationId: createSupplier
      requestBody:
        description: 'Required data for creation [uses SupplierCreateDTO]'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierCreateDTO'
      responses:
        '201':
          description: 'Supplier successfully created [returns SupplierDetailDTO]'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierDetailDTO'

  /suppliers/{id}:
    get:
      summary: Retrieve single supplier
      description: |
        Returns details for a supplier including cumulative rating stats
        and a list of all associated orders.
      operationId: getSupplierById
      parameters:
        - name: id
          in: path
          required: true
          description: The openBIS PermID of the supplier.
          schema:
            type: string
      responses:
        '200':
          description: 'Supplier found [returns SupplierDetailDTO]'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierDetailDTO'
        '404':
          description: Supplier not found
    put:
      summary: Edit supplier
      description: Updates master data of a supplier.
      operationId: updateSupplier
      parameters:
        - name: id
          in: path
          required: true
          description: The openBIS PermID.
          schema:
            type: string
      requestBody:
        description: 'Data to update [uses SupplierUpdateDTO]'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierUpdateDTO'
      responses:
        '200':
          description: 'Update successful [returns SupplierDetailDTO]'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierDetailDTO'

  # --------------------------------------------------------------------------
  # ORDERS
  # --------------------------------------------------------------------------
  /orders:
    get:
      summary: List orders
      description: Get all orders, optionally filtered by supplier.
      operationId: getOrders
      parameters:
        - name: supplierId
          in: query
          description: Filter by Supplier PermID
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 'List of orders [returns an array of OrderDetailDTO]'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDetailDTO'
    post:
      summary: Record new order
      description: Creates a new order (Type BESTELLUNG) and assigns it to a supplier. The code is generated automatically.
      operationId: createOrder
      requestBody:
        description: 'Required data for creation [uses OrderCreateDTO]'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateDTO'
      responses:
        '201':
          description: 'Order created [returns OrderDetailDTO]'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailDTO'

  /orders/{id}:
    get:
      summary: Retrieve order
      description: Returns details of an order including reference to the rating (if available).
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          required: true
          description: The openBIS PermID of the order.
          schema:
            type: string
      responses:
        '200':
          description: 'Order found [returns OrderDetailDTO]'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailDTO'
        '404':
          description: Order not found
    put:
      summary: Edit order
      description: Updates order data.
      operationId: updateOrder
      parameters:
        - name: id
          in: path
          required: true
          description: The openBIS PermID.
          schema:
            type: string
      requestBody:
        description: 'Data to update [uses OrderUpdateDTO]'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateDTO'
      responses:
        '200':
          description: 'Update successful [returns OrderDetailDTO]'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailDTO'

  # --------------------------------------------------------------------------
  # RATINGS
  # --------------------------------------------------------------------------
  /ratings:
    post:
      summary: Submit rating
      description: |
        Creates a rating (Type BESTELLBEWERTUNG) for an existing order.
        Calculates the total score automatically.
      operationId: createRating
      requestBody:
        description: 'Required data for creation [uses RatingCreateDTO]'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingCreateDTO'
      responses:
        '201':
          description: 'Rating created [returns RatingDetailDTO]'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RatingDetailDTO'

  /ratings/{id}:
    get:
      summary: Retrieve rating
      description: Returns details of a rating.
      operationId: getRatingById
      parameters:
        - name: id
          in: path
          required: true
          description: The openBIS PermID of the rating.
          schema:
            type: string
      responses:
        '200':
          description: 'Rating found [returns RatingDetailDTO]'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RatingDetailDTO'
        '404':
          description: Rating not found

components:
  schemas:
    # --- COMMON STATS ---
    RatingStats:
      title: RatingStats
      type: object
      description: Calculated average values for a supplier based on the 4 criteria and total score.
      properties:
        avgQuality:
          type: number
          format: float
          description: Average of QUALITAET
          examples:
            - 4.5
        avgCost:
          type: number
          format: float
          description: Average of KOSTEN
          examples:
            - 3.8
        avgReliability:
          type: number
          format: float
          description: Average of EINHALTUNG_TERMINE
          examples:
            - 5.0
        avgAvailability:
          type: number
          format: float
          description: Average of VERFUEGBARKEIT
          examples:
            - 4.2
        avgTotal:
          type: number
          format: float
          description: Average of GESAMTBEWERTUNG
          examples:
            - 4.4
        totalRatingCount:
          type: integer
          description: Number of ratings included in this calculation
          examples:
            - 12

    # --- SUPPLIER DTOS ---
    SupplierBaseDTO:
      title: SupplierBaseDTO
      type: object
      required:
        - name
        - customerNumber
        - street
        - country
        - zipCode
        - city
        - vatId
        - conditions
      properties:
        name:
          type: string
          description: Mapped to $NAME
          examples:
            - "Acme Corp"
        customerNumber:
          type: string
          description: Mapped to KUNDENNUMMER
          examples:
            - "KD-1234"
        addition:
          type: string
          description: Mapped to LIEFERANTEN_ZUSATZ
          examples:
            - "Building B"
        street:
          type: string
          description: Mapped to LIEFERANTEN_STRASSE
          examples:
            - "Industrial Ave 5"
        poBox:
          type: string
          description: Mapped to LIEFERANTEN_POSTFACH
        country:
          type: string
          description: Mapped to LIEFERANTEN_LAND (Vocabulary LIEFERANTEN_LAND)
          enum: [ CH, D, F, FL, NL ]
          examples:
            - "CH"
        zipCode:
          type: string
          description: Mapped to LIEFERANTEN_POSTLEITZAHL
          examples:
            - "8005"
        city:
          type: string
          description: Mapped to LIEFERANTEN_ORT
          examples:
            - "Zurich"
        website:
          type: string
          description: Mapped to LIEFERANTEN_WEBLINK
          pattern: '^(http|https|ftp)://\S+$'
          examples:
            - "https://acme.com"
        email:
          type: string
          description: Mapped to LIEFERANTEN_EMAIL
        phoneNumber:
          type: string
          description: Mapped to LIEFERANTEN_TELEFON
        vatId:
          type: string
          description: Mapped to MWST
        conditions:
          type: string
          description: Mapped to KONDITIONEN
        customerInfo:
          type: string
          description: Mapped to KUNDENINFORMATION

    SupplierCreateDTO:
      title: SupplierCreateDTO
      allOf:
        - $ref: '#/components/schemas/SupplierBaseDTO'

    SupplierUpdateDTO:
      title: SupplierUpdateDTO
      $ref: '#/components/schemas/SupplierBaseDTO'

    SupplierSummaryDTO:
      title: SupplierSummaryDTO
      allOf:
        - $ref: '#/components/schemas/SupplierBaseDTO'
        - type: object
          required:
            - id
            - code
            - stats
          properties:
            id:
              type: string
              description: OpenBIS PermID (Stable Identifier)
            code:
              type: string
              description: OpenBIS Code (Business Identifier)
            stats:
              $ref: '#/components/schemas/RatingStats'

    SupplierDetailDTO:
      title: SupplierDetailDTO
      allOf:
        - $ref: '#/components/schemas/SupplierSummaryDTO'
        - type: object
          properties:
            orders:
              type: array
              description: List of all orders belonging to this supplier
              items:
                $ref: '#/components/schemas/OrderDetailDTO'

    # --- ORDER DTOS ---
    OrderBaseDTO:
      title: OrderBaseDTO
      type: object
      required:
        - name
        - mainCategory
        - subCategory
        - reason
        - orderedBy
        - orderDate
      properties:
        name:
          type: string
          description: Mapped to NAME (Short description)
          examples:
            - "Lab material Q1"
        mainCategory:
          type: string
          description: Mapped to BESTELLUNG_HK (Vocabulary BESTELLUNGHK)
          enum: [ Beschaffung, Dienstleistung ]
          examples:
            - "Beschaffung"
        subCategory:
          type: string
          description: Mapped to BESTELLUNG_UK (Vocabulary BESTELLUNGFK)
          enum:
            - Beratung
            - Dienstleistung
            - Gerät/Werkzeug
            - Maschine
            - Messgeräte
            - Messmittel
            - PC Hardware
            - PC Software
            - Prüfmaschine
          examples:
            - "Messmittel"
        details:
          type: string
          description: Mapped to BEZEICHNUNG
          examples:
            - "50x Precision scales"
        frequency:
          type: string
          description: Mapped to RYTHMUS [sic!]
          examples:
            - "One-time"
        contactPerson:
          type: string
          description: Mapped to ANSPRECHPERSON
          examples:
            - "John Doe"
        contactEmail:
          type: string
          description: Mapped to ANSPRECHPERSON_EMAIL
        contactPhone:
          type: string
          description: Mapped to ANSPRECHPERSON_TELEFON
        reason:
          type: string
          description: Mapped to BESCHAFFUNGSGRUND
        orderMethod:
          type: string
          description: Mapped to BESTELLART
          examples:
            - "Email"
        orderedBy:
          type: string
          description: Mapped to BESTELLER (Name of initiator)
        orderDate:
          type: string
          format: date
          description: Mapped to BESTELLDATUM (YYYY-MM-DD)
          examples:
            - "2023-12-15"
        deliveryDate:
          type: string
          format: date
          description: Mapped to LIEFERDATUM (YYYY-MM-DD)
        orderComment:
          type: string
          description: Mapped to KOMMENTAR (General comment)

    OrderCreateDTO:
      title: OrderCreateDTO
      allOf:
        - $ref: '#/components/schemas/OrderBaseDTO'
        - type: object
          required:
            - supplierId
          properties:
            supplierId:
              type: string
              description: Reference to supplier (PermID) - mandatory for assignment
              examples:
                - "2023050512345"

    OrderUpdateDTO:
      title: OrderUpdateDTO
      $ref: '#/components/schemas/OrderBaseDTO'

    OrderDetailDTO:
      title: OrderDetailDTO
      allOf:
        - $ref: '#/components/schemas/OrderBaseDTO'
        - type: object
          required:
            - id
            - code
          properties:
            id:
              type: string
              description: OpenBIS PermID
            code:
              type: string
              description: OpenBIS Code
            ratingStatus:
              type: string
              enum: [ RATED, PENDING ]
              description: Helps the frontend to show if already rated.
            supplierId:
              type: string
              description: PermID of the parent supplier
            supplierName:
              type: string
              description: Name of the parent supplier (for display)
            ratingId:
              type: [ "string", "null" ]
              description: PermID of the linked rating (if exists), otherwise null.

    # --- RATING DTOS ---
    RatingBaseDTO:
      title: RatingBaseDTO
      type: object
      required:
        - quality
        - qualityReason
        - cost
        - costReason
        - reliability
        - reliabilityReason
      properties:
        quality:
          type: integer
          minimum: 1
          maximum: 5
          format: integer
          description: Mapped to QUALITAET (1-5), 1 = Poor, 5 = Excellent
          examples:
            - 5
        qualityReason:
          type: string
          description: Mapped to GRUND_QUALITAET
        cost:
          type: integer
          minimum: 1
          maximum: 5
          format: integer
          description: Mapped to KOSTEN (1-5), 1 = Poor, 5 = Excellent
          examples:
            - 3
        costReason:
          type: string
          description: Mapped to KOSTEN_BEGRUENDUNG
        reliability:
          type: integer
          minimum: 1
          maximum: 5
          format: integer
          description: Mapped to EINHALTUNG_TERMINE (1-5), 1 = Poor, 5 = Excellent
          examples:
            - 5
        reliabilityReason:
          type: string
          description: Mapped to EINHALTUNG_TERMINE_GRUND
        availability:
          type: integer
          minimum: 1
          maximum: 5
          format: integer
          description: Mapped to VERFUEGBARKEIT (1-5) 1 = Poor, 5 = Excellent
          examples:
            - 4
        availabilityReason:
          type: string
          description: Mapped to VERFUEGBARKEIT_BEGRUENDUNG
        ratingComment:
          type: string
          description: Mapped to KOMMENTAR

    RatingCreateDTO:
      title: RatingCreateDTO
      allOf:
        - $ref: '#/components/schemas/RatingBaseDTO'
        - type: object
          required:
            - orderId
            - quality
            - qualityReason
            - cost
            - costReason
            - reliability
            - reliabilityReason
          properties:
            orderId:
              type: string
              description: PermID of the order this rating belongs to.

    RatingDetailDTO:
      title: RatingDetailDTO
      allOf:
        - $ref: '#/components/schemas/RatingBaseDTO'
        - type: object
          required:
            - id
            - code
            - orderId
            - totalScore
          properties:
            id:
              type: string
              description: PermID of the rating
            code:
              type: string
              description: OpenBIS Code of the rating
            orderId:
              type: string
              description: PermID of the order this rating belongs to.
            totalScore:
              type: number
              format: float
              description: Mapped to GESAMTBEWERTUNG (Calculated by Backend)
              examples:
                - 4.4
            supplierId:
              type: string
              description: For navigation to supplier
            supplierName:
              type: string